+++ src/thrustsystem.cpp
@@ -14,7 +14,7 @@ void ThrustSystem::update()
 		if (!thrust || !state)
 			continue;
 
-		double thrust_multiplier = 1.0;
+		float thrust_multiplier = 1.0;
 		if ((state->current.vel.x > 0.0 && thrust->current.x < 0.0) || (state->current.vel.x < 0.0 && thrust->current.x > 0.0))
 			thrust_multiplier = thrust->reverse_thrust_factor;
 		thrust->current.x *= thrust_multiplier;
